+++ src/exactly_lib/definitions/os_proc_env.py
@@ -7,8 +7,8 @@ ENV_VARS__DEFAULT = None
 ENV_VARS_GETTER__DEFAULT = os_environ_getter
 TIMEOUT__DEFAULT = None
 
-TIMEOUT__NONE_STRING = 'none'
+TIMEOUT__NONE_TOKEN = 'none'
 
 
 def render_timeout_value(x: Optional[int]) -> str:
-    return functional.reduce_optional(str, TIMEOUT__NONE_STRING, x)
+    return functional.reduce_optional(str, TIMEOUT__NONE_TOKEN, x)