+++ src/exactly_lib/impls/instructions/multi_phase/timeout/defs.py
@@ -1,3 +1,4 @@
-from exactly_lib.definitions import instruction_arguments
+from exactly_lib.definitions import instruction_arguments, os_proc_env
 
 ASSIGNMENT_IDENTIFIER = instruction_arguments.ASSIGNMENT_OPERATOR
+NONE_TOKEN = os_proc_env.TIMEOUT__NONE_TOKEN